+++ src/main/java/org/sid/secservice/sec/rep/MainOeuvreRepository.java
@@ -3,5 +3,5 @@ package org.sid.secservice.sec.rep;
 import org.sid.secservice.sec.entities.MainOeuvre;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MainOeuvreRepository extends JpaRepository<MainOeuvre,Integer> {
+public interface MainOeuvreRepository extends JpaRepository<MainOeuvre,Long> {
 }
