+++ src/main/java/org/sid/secservice/sec/rep/PieceRepository.java
@@ -3,5 +3,5 @@ package org.sid.secservice.sec.rep;
 import org.sid.secservice.sec.entities.Piece;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PieceRepository extends JpaRepository<Piece,Integer> {
+public interface PieceRepository extends JpaRepository<Piece,Long> {
 }
